Fix build for Xcode 11.4.1: - Update deployment target to iOS 8 and change std c++ lib to default compiler settings to fix issues with ibstdc++ (building with libstdc++ was deprecated with Xcode 8 and is not supported in Xcode 10 when targeting iOS). - Fix types conversions where it is necessary. - Call super methods implementations where it is necessary. - Use correct includes for new std lib. - Use new API instead of deprecated where it is necessary (use quick solutions). - Remove missed resources, anyway they are not accessible. - Add base.iPack and doom.wad.

+++ common/embeddedaudiosynthesis/EASGlue.c
@@ -37,8 +37,8 @@ static EAS_BOOL EASLibraryCheck (const S_EAS_LIB_CONFIG *pLibConfig);
 
 static EAS_DATA_HANDLE pEASData;
 static const S_EAS_LIB_CONFIG *pLibConfig;
-static int polyphony;
-static int bufferSize;
+static long polyphony;
+static long bufferSize;
 
 static EAS_FILE file;
 static EAS_HANDLE handle;
